LayoutTests/imported/w3c/web-platform-tests/fetch/api/basic/integrity-sharedworker-expected.txt

CONSOLE MESSAGE: line 12: ReferenceError: Can't find variable: SharedWorker

FAIL Fetch in sharedworker: integrity handling ReferenceError: Can't find variable: SharedWorker
